fixing connection refuse

+++ app/src/main/java/com/dodolife/jadwalsholat/database/repo/PrayerTimesRepo.kt
@@ -6,6 +6,7 @@ import com.dodolife.jadwalsholat.database.toDb
 import com.dodolife.jadwalsholat.network.service.PrayerTimesService
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import java.text.DecimalFormat
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -18,11 +19,10 @@ class PrayerTimesRepo @Inject constructor(
     private val prayerTimesDao: PrayerTimesDao = jadwalDb.PrayerTimesDao()
 
     suspend fun getPrayerTimes(
-        latitude: Long,
-        longitude: Long
+        latitude: Double,
+        longitude: Double
     ) = withContext(Dispatchers.IO) {
         val response = prayerTimesService.getPrayerTimes(latitude,longitude).await()
         prayerTimesDao.insert(response.data.map { it.toDb() })
     }
-
 }
